# docs/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "StudyAss 学习助手 API",
    "description": "StudyAss AI 学习平台 - RESTful API 接口文档",
    "version": "1.0.0",
    "contact": {
      "name": "俊哥的学习助手团队"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "本地开发环境"
    },
    {
      "url": "https://api.studyass.com/api",
      "description": "生产环境"
    }
  ],
...
  "paths": {
...
    "/auth/send-code": {
      "post": {
        "tags": ["用户认证"],
        "summary": "发送验证码",
        "operationId": "sendCode",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendCodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "验证码已发送"
          },
          "400": {
            "description": "请求参数错误"
          },
          "429": {
            "description": "发送过于频繁"
          }
        }
      }
    },
...
    "/knowledge": {
      "get": {
        "tags": ["知识点管理"],
        "summary": "获取知识点列表",
        "operationId": "getKnowledgeList",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {"type": "integer", "default": 1}
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {"type": "integer", "default": 20}
          },
          {
            "name": "subject",
            "in": "query",
            "schema": {"type": "string"}
          },
          {
            "name": "grade",
            "in": "query",
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "知识点列表"
          }
        }
      },
      "post": {
        "tags": ["知识点管理"],
        "summary": "创建知识点",
        "operationId": "createKnowledge",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功"
          }
        }
      }
    },
...
    "/progress": {
      "get": {
        "tags": ["学习进度"],
        "summary": "获取学习进度列表",
        "operationId": "getProgressList",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "进度列表"
          }
        }
      }
    },
...
    "/ai/planning/generate": {
      "post": {
        "tags": ["AI 学习规划"],
        "summary": "生成学习计划",
        "operationId": "generatePlan",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratePlanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "计划生成成功"
          }
        }
      }
    },
    "/ai/planning/user-profile": {
      "get": {
        "tags": ["AI 学习规划"],
        "summary": "获取用户学习画像",
        "operationId": "getUserProfile",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "subject",
            "in": "query",
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "用户画像"
          }
        }
      }
    },
    "/ai/planning/daily-tasks/{date}": {
      "get": {
        "tags": ["AI 学习规划"],
        "summary": "获取每日任务",
        "operationId": "getDailyTasks",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "date"}
          }
        ],
        "responses": {
          "200": {
            "description": "每日任务列表"
          }
        }
      }
    },
    "/textbooks/upload": {
      "post": {
        "tags": ["课本解析"],
        "summary": "上传课本 PDF",
        "operationId": "uploadTextbook",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/TextbookUploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "上传成功"
          }
        }
      }
    },
    "/textbooks": {
      "get": {
        "tags": ["课本解析"],
        "summary": "获取课本列表",
        "operationId": "getTextbooks",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "课本列表"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "SendCodeRequest": {
        "type": "object",
        "required": ["phone"],
        "properties": {
          "phone": {"type": "string", "example": "13800138000"},
          "purpose": {"type": "string", "enum": ["login", "register"], "default": "login"}
        }
      },
...
      "KnowledgeRequest": {
        "type": "object",
        "properties": {
          "name": {"type": "string"},
          "subject": {"type": "string"},
          "grade": {"type": "integer"},
          "description": {"type": "string"},
          "parent_id": {"type": "integer", "nullable": true}
        }
      },
...
      "GeneratePlanRequest": {
        "type": "object",
        "properties": {
          "subject": {"type": "string"},
          "timeframe": {
            "type": "object",
            "properties": {
              "total_days": {"type": "integer"},
              "start_date": {"type": "string", "format": "date"},
              "end_date": {"type": "string", "format": "date"}
            }
          },
          "goals": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "knowledge_point_id": {"type": "integer"},
                "current_mastery": {"type": "number"},
                "target_mastery": {"type": "number"},
                "priority": {"type": "string", "enum": ["low", "medium", "high"]}
              }
            }
          },
          "preferences": {
            "type": "object",
            "properties": {
              "daily_time_limit": {"type": "integer"},
              "include_weekend": {"type": "boolean"},
              "learning_style": {"type": "string"}
            }
          }
        }
      },
      "TextbookUploadRequest": {
        "type": "object",
        "properties": {
          "file": {"type": "string", "format": "binary"},
          "title": {"type": "string"},
          "subject": {"type": "string"},
          "grade": {"type": "string"},
          "version": {"type": "string"}
        }
      },
...
    }
  }
}